Functioning friction and a car

+++ SDL_GameEngine/FluidSimulation.cpp
@@ -7,21 +7,21 @@
 #include <random>
 #include "SDL2_gfxPrimitives.h"
 
-const float smoothingRadius = 200;
-const float mass = 1;
+const double smoothingRadius = 200;
+const double mass = 1;
 
-float SmoothingKernel(float radius, float dst)
+double SmoothingKernel(double radius, double dst)
 {
 	if (dst >= radius) return 0;
 
-	float volume = (M_PI * pow(radius, 4)) / 6;
+	double volume = (M_PI * pow(radius, 4)) / 6;
 	return (radius - dst) * (radius - dst) / volume;
 }
-float SmoothingKernelDerivative(float dst, float radius)
+double SmoothingKernelDerivative(double dst, double radius)
 {
 	if (dst >= radius) return 0;
 
-	float scale = 12 / (M_PI * std::pow(radius, 4));
+	double scale = 12 / (M_PI * std::pow(radius, 4));
 	return (dst - radius) * scale;
 }
 
@@ -34,15 +34,15 @@ void FluidSimulation_System::Init(Registry* reg)
 			Fluid_Component* fluid = &reg->fluids[e];
 			int particlesPerRow = (int)std::sqrt(fluid->particleCount);
 			int particlesPerCol = (fluid->particleCount - 1) / particlesPerRow + 1;
-			float spacing = fluid->particleRadius * 2 + fluid->startSpacing;
+			double spacing = fluid->particleRadius * 2 + fluid->startSpacing;
 
 			std::default_random_engine generator;
 			std::uniform_int_distribution<int> distribution(960, 2900);
 			std::uniform_int_distribution<int> distribution2(120, 2000);
 			for (int i = 0; i < reg->fluids[e].particleCount; i++)
 			{
-				float x = (i % particlesPerRow - particlesPerRow / 2.f + 0.5f) * spacing;
-				float y = (i / particlesPerRow - particlesPerCol / 2.f + 0.5f) * spacing;
+				double x = (i % particlesPerRow - particlesPerRow / 2.f + 0.5f) * spacing;
+				double y = (i / particlesPerRow - particlesPerCol / 2.f + 0.5f) * spacing;
 
 				std::cout << "REACHED";
 				fluid->positions.push_back(Vector2D(x + fluid->startPos.x, y + fluid->startPos.y));
@@ -79,10 +79,10 @@ void FluidSimulation_System::Update(double* deltaTime)
 
 void FluidSimulation_System::ResolveCollision(int i)
 {
-	float down = fluid->positions[i].y + fluid->particleRadius;
-	float up = fluid->positions[i].y - fluid->particleRadius;
-	float right = fluid->positions[i].x + fluid->particleRadius;
-	float left = fluid->positions[i].x - fluid->particleRadius;
+	double down = fluid->positions[i].y + fluid->particleRadius;
+	double up = fluid->positions[i].y - fluid->particleRadius;
+	double right = fluid->positions[i].x + fluid->particleRadius;
+	double left = fluid->positions[i].x - fluid->particleRadius;
 
 	if (down > fluid->boundingBox.y + fluid->boundingBox.h)
 	{
@@ -106,14 +106,14 @@ void FluidSimulation_System::ResolveCollision(int i)
 	}
 }
 
-float FluidSimulation_System::CalculateDensity(Vector2D samplePoint)
+double FluidSimulation_System::CalculateDensity(Vector2D samplePoint)
 {
-	float density = 0;
+	double density = 0;
 
 	for (Vector2D position : fluid->positions)
 	{
-		float dst = (position - samplePoint).length();
-		float influence = SmoothingKernel(smoothingRadius, dst);
+		double dst = (position - samplePoint).length();
+		double influence = SmoothingKernel(smoothingRadius, dst);
 		density += mass * influence;
 	}
 
@@ -129,21 +129,21 @@ Vector2D FluidSimulation_System::CalculatePressureForce(int index)
 		if (index == i) continue;
 
 		Vector2D offset = fluid->positions[i] - fluid->positions[index];
-		float dst = offset.length();
+		double dst = offset.length();
 		Vector2D dir = dst == 0 ? Vector2D(1, 1) : offset / dst;
 
-		float slope = SmoothingKernelDerivative(dst, smoothingRadius);
-		float density = fluid->densities[i];
-		float sharedPressure = (ConvertDensityToPressure(density) + ConvertDensityToPressure(fluid->densities[index])) / 2;
+		double slope = SmoothingKernelDerivative(dst, smoothingRadius);
+		double density = fluid->densities[i];
+		double sharedPressure = (ConvertDensityToPressure(density) + ConvertDensityToPressure(fluid->densities[index])) / 2;
 		pressureForce += dir * -sharedPressure * slope * mass / density;
 	}
 
 	return pressureForce;
 }
-float FluidSimulation_System::ConvertDensityToPressure(float density)
+double FluidSimulation_System::ConvertDensityToPressure(double density)
 {
-	float densityError = density - fluid->targetDensity;
-	float pressure = densityError * fluid->pressureMultiplier;
+	double densityError = density - fluid->targetDensity;
+	double pressure = densityError * fluid->pressureMultiplier;
 	return pressure;
 }
