+++ SDL_GameEngine/Game.h
@@ -29,8 +29,8 @@ public:
 	const int rectangleLineThickness = 10;
 	const SDL_Color backroundColor = SDL_Color{ 214, 206, 194, 255 };
 	const SDL_Rect rect = { 920 / rectangleLineThickness, 80 / rectangleLineThickness , 2000 / rectangleLineThickness , 2000 / rectangleLineThickness };
-	float* gravity;
-	float* dragCoeficient;
+	double* gravity;
+	double* dragCoeficient;
 	std::string assetPath = "";
 	bool stopSimulation = false;
 private:
